Beginning of splitting out goomba-specific stuff from gba emulator base class
